pyproject.toml
[tool.poetry]
name = "fmc-backend"
version = "0.1.0"
description = ""
authors = ["Varvara"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
requests = "^2.28.2"
uvicorn = "^0.20.0"
fastapi-pagination = "^0.11.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^1.10.5"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.4"
email-validator = "^1.3.1"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.4"
ipython = "^8.11.0"
sqladmin = "^0.9.0"
typer = "^0.7.0"
typer-cli = "^0.0.13"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
